+++ CHANGELOG.md
@@ -1,7 +1,11 @@
-## 2.0.4 (Unreleased)
+## 2.1.0 (Unreleased)
 
 DOCS:
 * Fixed usage example for sumologic_content resource.
+
+DEPRECATIONS:
+* resource/sumologic_collector: Remove deprecated attributes `lookup_by_name` and `destroy` (#https://github.com/terraform-providers/terraform-provider-sumologic/pull/32)
+* resource/sumologic_sources: Remove deprecated attributes `lookup_by_name` and `destroy` (#https://github.com/terraform-providers/terraform-provider-sumologic/pull/32)
 
 
 ## 2.0.3 (June 02, 2020)